440fx: Optimize static PRT generation
There is an exisitng function build_prt, which has been further
optimized by a new function. This new function was introduced for
q35. Now it has been renamed and being used for both platforms.
This speeds up the boot time, as the generated AML table is only
dummy one. Reduces the boot time parsing of AML by the kernel.

Signed-off-by: Arindam Roy <arindam.roy@intel.com>